+++ src/Controllers/Live/Calls.php
@@ -1,32 +1,37 @@
 <?php
 
-namespace WombatDialer\Reports;
+namespace WombatDialer\Controllers\Live;
 
 use Illuminate\Support\Facades\Http;
 use WombatDialer\Controllers\Edit\Wombat;
 
-class Logs extends Wombat
+class Calls extends Wombat
 {
     //protected $path = '/edit/asterisk';
     use \WombatDialer\Concerns\LiveCallsandReportsTraits;
 
     /**
      * Perform API GET.
-     * Returns the Log set for the set of runs.
+     * Returns the list of all the Live calls.
      *
-     * @params  $items and $from and $id
      * @return json response
      */
-    public function reportLogs($from, $items, $id)
+    public function liveCalls()
     {
-        $this->path = '/reports/logs/';
-        $from = $from ? $from : null;
-        $items = $items ? $items : null;
-        $id = $id ? $id : null;
-        $this->query = ['from' => $from, 'items' => $items, 'id' => $id];
         $response = Http::withBasicAuth($this->userAuth(), $this->passAuth())
             ->get($this->connection());
 
         return $response->json();
+    }
+
+    /**
+     * Perform API GET.
+     * Returns the list of all the Live calls.
+     *
+     * @return json response
+     */
+    public function showliveCalls()
+    {
+        return $this->liveCalls();
     }
 }
